editing a bucket feature
