+++ computerapi/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
-from .views import read_all_bucket, add_bucket
+from .views import read_all_bucket, add_bucket ,edit_bucket
 
 urlpatterns = [
     path('api/v.1/allbuckets', read_all_bucket, name ='read_all_buckets'),
+    path('api/v.1/editbucket/<bucket_id>', edit_bucket, name ='edit_bucket'),
     path('api/v.1/addbucket/', add_bucket, name ='add_bucket'),
 ]
