+++ computerapi/views/__init__.py
@@ -1,1 +1,3 @@
-from .views import read_all_bucket, add_bucket
+from .views import read_all_bucket
+from .edit_bucket_view import edit_bucket
+from .add_bucket_view import add_bucket
