+++ computerapi/views/views.py
@@ -8,17 +8,16 @@ from django.views.decorators.csrf import csrf_exempt
 from rest_framework.decorators import api_view
 
 # Create your views here
-@api_view(['POST'])
-@csrf_exempt
-def add_bucket(request):
-    if request.method == 'POST':
-        data = request.data
-        print(data)
-        serializer = BucketSerializer(data = data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status = status.HTTP_201_CREATED)
-        return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+# @api_view(['POST'])
+# # @csrf_exempt
+# def add_bucket(request):
+#     if request.method == 'POST':
+#         data = request.data
+#         serializer = BucketSerializer(data = data)
+#         if serializer.is_valid():
+#             serializer.save()
+#             return Response(serializer.data, status = status.HTTP_201_CREATED)
+#         return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET'])
 def read_all_bucket(request):
@@ -26,10 +25,3 @@ def read_all_bucket(request):
         all_buckets = Bucket.objects.all()
         serializer = BucketSerializer(all_buckets, many = True)
         return Response(serializer.data)
-
-def update_bucket(request, bucket_id):
-    if request.method == 'POST':
-        all_buckets = Bucket.objects.get(pk = bucket_id)
-        all_buckets.name = request.data['name']
-        serializer = BucketSerializer(all_buckets, many = True)
-        return JsonResponse(serializer.data, safe = False)
